+++ app/Modules/Clients/Views/clients/_modal_edit.blade.php
@@ -5,10 +5,10 @@
     <div class="modal-dialog modal-xl">
         <div class="modal-content">
             <div class="modal-header">
-                <h4 class="modal-title">@lang('bt.edit_client')</h4>
-                <div class="float-end">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">@lang('bt.cancel')</button>
-                    <input type="submit" id="btn-edit-client-submit" class="btn btn-primary" value="@lang('bt.save')">
+                <div class="modal-title col-12">
+                <h4 class="ms-3 float-start">@lang('bt.edit_client')</h4>
+                    <button type="button" class="btn btn-secondary me-3 float-end" data-bs-dismiss="modal">@lang('bt.cancel')</button>
+                    <input type="submit" id="btn-edit-client-submit" class="btn btn-primary me-3 float-end" value="@lang('bt.save')">
                 </div>
             </div>
             <div class="modal-body">