+++ app/Modules/Vendors/Models/Vendor.php
@@ -136,6 +136,10 @@ class Vendor extends Model
     | Accessors
     |--------------------------------------------------------------------------
     */
+    public function isFlagged(): bool
+    {
+        return $this->flag_text != null;
+    }
 
     public function attachmentPath(): Attribute
     {